fastapi/requirements.txt
alembic==1.9.2
anyio==3.6.2
boto3==1.24.46
botocore==1.27.59
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.4
docker==6.0.1
entrypoints==0.4
fastapi==0.88.0
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mlflow==2.0.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.3
Pillow==9.4.0
protobuf==4.21.12
pyarrow==10.0.1
pydantic==1.10.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
starlette==0.22.0
tabulate==0.9.0
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
websocket-client==1.5.0
Werkzeug==2.2.2
zipp==3.12.0
